// src/mysql/dto/user.dto.ts
/**
 * ```ts
 * interface userDTO{
 * id: number
 * username: string
 * email: string
 * image: string
 * token: null | string
 * }
 * ```
*/

// export interface userDTO{
//   username: string
//   email: string
//   image?: string
// }

export interface userDTO{
  id: number
  username: string
  email: string
  image: string
  token: null | string
}